// src/interfaces.ts
export enum ChartType {
	SIMPLE_BAR_CHART = 'simple-bar-chart',
	SCATTER_CHART = 'scatter-chart',
	DONUT_CHART = 'donut-chart',
	STACKED_BAR_CHART = 'stacked-bar-chart',
	LINE_CHART = 'line-chart',
	AREA_CHART = 'area-chart',
	STACKED_AREA_CHART = 'stacked-area-chart',
	HORIZONTAL_BAR_CHART = 'horizontal-bar-chart',
	BUBBLE_CHART = 'bubble-chart',
	GROUPED_BAR_CHART = 'grouped-bar-chart',
	PIE_CHART = 'pie-chart'
}